+++ packages/form-builder-react/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@buildnbuzz/form-builder-react",
-  "version": "0.1.0",
+  "version": "0.0.0",
   "description": "React adapter for the BuzzForm form builder.",
   "author": "Parth Lad <parth@buildnbuzz.com>",
   "license": "MIT",
@@ -86,6 +86,6 @@
   },
   "publishConfig": {
     "access": "public",
-    "provenance": true
+    "provenance": false
   }
 }
