+++ packages/form-builder-react/src/components/draggable-sidebar-item.tsx
