+++ packages/form-builder-react/src/components/sortable-node.tsx
