+++ packages/form-builder-react/src/hooks/index.ts
@@ -1,4 +1,4 @@
-export * from "./useDropIndicator";
+export * from "./use-drop-indicator";
 export * from "./use-auto-save";
 export * from "./use-sync-property";
 export * from "./use-keyboard-shortcuts";
