+++ packages/form-builder-react/src/index.ts
@@ -1,20 +1,20 @@
 export * from "./types";
 export * from "./registry";
 export * from "./state/store";
-export * from "./context/BuilderContext";
-export * from "./components/BuilderCanvas";
-export * from "./components/BuilderNode";
-export * from "./components/BuilderSidebar";
-export * from "./components/BuilderProperties";
-export * from "./components/BuilderDndProvider";
-export * from "./components/BuilderFormProvider";
-export * from "./components/BuilderFormStateSync";
-export * from "./components/SortableNode";
-export * from "./components/DraggableSidebarItem";
+export * from "./context/builder-context";
+export * from "./components/builder-canvas";
+export * from "./components/builder-node";
+export * from "./components/builder-sidebar";
+export * from "./components/builder-properties";
+export * from "./components/builder-dnd-provider";
+export * from "./components/builder-form-provider";
+export * from "./components/builder-form-state-sync";
+export * from "./components/sortable-node";
+export * from "./components/draggable-sidebar-item";
 export { DEFAULT_SLOT } from "@buildnbuzz/form-builder-core";
 export * from "./hooks";
 export * from "./persistence/local-storage-provider";
 export { downloadTextFile } from "./utils/download";
 export { createExpressionStore } from "./state/expression-store";
 export type { ExpressionStoreState } from "./state/expression-store";
-export { ExpressionProvider, useExpressionContext, useAvailableFields } from "./context/ExpressionContext";
+export { ExpressionProvider, useExpressionContext, useAvailableFields } from "./context/expression-context";
